:recycle: refactor: error 폴더 ErrorBoundary로 이름 변경
