+++ frontend/src/components/common/ErrorBoundary/AuthError.ts
