+++ frontend/src/components/common/ErrorBoundary/AuthErrorBoundary.tsx
